[Twinkle#67#2] Refactor Direct SQL Calls to jOOQ Queries

+++ twinkle-app/twinkle.dal.abstractions/src/main/java/com/github/donkeyrit/twinkle/dal/common/repositories/GenericRepository.java
@@ -5,7 +5,7 @@ import java.util.stream.Stream;
 
 public interface GenericRepository<T extends Identifiable> 
 {
-	T findById(Long id);
+	T findById(int id);
 	Stream<T> findAll();
 	boolean save(T o);
 	boolean delete(T o);
