+++ twinkle-app/twinkle.dal.abstractions/src/main/java/com/github/donkeyrit/twinkle/dal/models/Rent.java
@@ -6,7 +6,7 @@ import com.github.donkeyrit.twinkle.dal.common.models.Identifiable;
 
 public class Rent implements Identifiable {
 
-    private Long id;
+    private int id;
     private int idClient;
     private int idCar;
     private Date startDate;
@@ -17,7 +17,7 @@ public class Rent implements Identifiable {
     public Rent() {
     }
 
-    public Rent(Long id, int idClient, int idCar, Date startDate, Date planDate, Date endDate) {
+    public Rent(int id, int idClient, int idCar, Date startDate, Date planDate, Date endDate) {
         this.id = id;
         this.idClient = idClient;
         this.idCar = idCar;
@@ -27,11 +27,11 @@ public class Rent implements Identifiable {
     }
 
     // Getters and Setters
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
